Feature: Embedded with online help, so that user may always get the most up-to-date help. Bugfix:  Use long instead of int for volume. This is just a partial fix. We still need to do it for the rest.

+++ src/org/yccheok/jstock/gui/IndicatorTableModel.java
@@ -229,7 +229,8 @@ public class IndicatorTableModel extends AbstractTableModelWithMemory {
         Double.class,
         Double.class,
         Double.class,
-        Integer.class,
+        // TODO: CRITICAL LONG BUG REVISED NEEDED.
+        Long.class,
         Double.class,
         Double.class,
         Integer.class,
